renamed PublicSuffixes2 to PublicSuffixes. deleted debug println.
